requirements.txt

requests
streamlit
pandas
Pillow
